Lista de correções
- agora voce pode publicar um site
- ao publicar aparece na galeria
- ao clicar em um portfolio é possivel ter uma visualizacao publica
- criada slug /portfolio/{subdomain}
- ao fazer login, carrega sua preview de portfolio, se existir uma

+++ src/config/firebase.js
@@ -1,8 +1,7 @@
-// src/config/firebase.js
 import { initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
-import { getFirestore, collection, getDocs, doc, getDoc, setDoc, deleteDoc, query, where } from 'firebase/firestore';
-import { getStorage, ref, uploadBytes, getDownloadURL } from 'firebase/storage';
+import { getFirestore, collection, query, where, getDocs, doc, getDoc, setDoc, deleteDoc } from 'firebase/firestore';
+import { getStorage } from 'firebase/storage';
 import firebaseConfig from './serviceAccountKey.js';
 
 const app = initializeApp(firebaseConfig);
@@ -10,19 +9,4 @@ const auth = getAuth(app);
 const db = getFirestore(app);
 const storage = getStorage(app);
 
-export { 
-  auth, 
-  db, 
-  storage, 
-  collection, 
-  getDocs, 
-  doc, 
-  getDoc, 
-  setDoc, 
-  deleteDoc, 
-  ref, 
-  uploadBytes, 
-  getDownloadURL,
-  query,
-  where
-};
+export { auth, db, storage, collection, query, where, getDocs, doc, getDoc, setDoc, deleteDoc };
